Change fragment generation target
Using the fragment name as the makefile target created a dependency loop
that could result in regenerating the fragment when you actually
intended to regenerate the config. Changed the target to a non-filename
and not a dependency of the config target to genfragment to ensure that
it will only be called intentionally.